+++ Xplat/src/main/java/vazkii/botania/common/item/block/SpecialFlowerBlockItem.java
@@ -9,13 +9,12 @@
 package vazkii.botania.common.item.block;
 
 import net.minecraft.ChatFormatting;
+import net.minecraft.locale.Language;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.tags.TagKey;
 import net.minecraft.util.Mth;
 import net.minecraft.world.item.BlockItem;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
@@ -30,9 +29,6 @@ import vazkii.botania.xplat.BotaniaConfig;
 import java.util.List;
 
 public class SpecialFlowerBlockItem extends BlockItem {
-	private static final TagKey<Item> GENERATING = BotaniaTags.Items.GENERATING_SPECIAL_FLOWERS;
-	private static final TagKey<Item> FUNCTIONAL = BotaniaTags.Items.FUNCTIONAL_SPECIAL_FLOWERS;
-	private static final TagKey<Item> MISC = BotaniaTags.Items.MISC_SPECIAL_FLOWERS;
 
 	public SpecialFlowerBlockItem(Block block1, Properties props) {
 		super(block1, props);
@@ -42,18 +38,22 @@ public class SpecialFlowerBlockItem extends BlockItem {
 	public void appendHoverText(@NotNull ItemStack stack, Level world, @NotNull List<Component> tooltip, @NotNull TooltipFlag flag) {
 		// Prevent crash when tooltips queried before configs load
 		if (BotaniaConfig.client() != null) {
-			if (world != null) {
-				if (stack.is(GENERATING)) {
-					tooltip.add(Component.translatable("botania.flowerType.generating").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
-				} else if (stack.is(FUNCTIONAL)) {
-					tooltip.add(Component.translatable("botania.flowerType.functional").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
-				} else if (stack.is(MISC)) {
-					tooltip.add(Component.translatable("botania.flowerType.misc").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
-				}
+			if (stack.is(BotaniaTags.Items.GENERATING_SPECIAL_FLOWERS)
+					|| stack.is(BotaniaTags.Items.GENERATING_FLOATING_FLOWERS)) {
+				tooltip.add(Component.translatable("botania.flowerType.generating").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
+			} else if (stack.is(BotaniaTags.Items.FUNCTIONAL_SPECIAL_FLOWERS)
+					|| stack.is(BotaniaTags.Items.FUNCTIONAL_FLOATING_FLOWERS)) {
+				tooltip.add(Component.translatable("botania.flowerType.functional").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
+			} else if (stack.is(BotaniaTags.Items.MISC_SPECIAL_FLOWERS)
+					|| stack.is(BotaniaTags.Items.MISC_FLOATING_FLOWERS)) {
+				tooltip.add(Component.translatable("botania.flowerType.misc").withStyle(ChatFormatting.ITALIC, ChatFormatting.BLUE));
 			}
 
 			if (BotaniaConfig.client().referencesEnabled()) {
 				String key = getDescriptionId() + ".reference";
+				if (key.contains("floating_") && !Language.getInstance().has(key)) {
+					key = key.replace("floating_", "");
+				}
 				MutableComponent lore = Component.translatable(key);
 				if (!lore.getString().equals(key)) {
 					tooltip.add(lore.withStyle(ChatFormatting.ITALIC, ChatFormatting.GRAY));