+++ Xplat/src/main/java/vazkii/botania/data/ItemTagProvider.java
@@ -65,6 +65,9 @@ public class ItemTagProvider extends ItemTagsProvider {
 		this.copy(BotaniaTags.Blocks.MISC_SPECIAL_FLOWERS, BotaniaTags.Items.MISC_SPECIAL_FLOWERS);
 		this.copy(BotaniaTags.Blocks.GENERATING_SPECIAL_FLOWERS, BotaniaTags.Items.GENERATING_SPECIAL_FLOWERS);
 		this.copy(BotaniaTags.Blocks.FUNCTIONAL_SPECIAL_FLOWERS, BotaniaTags.Items.FUNCTIONAL_SPECIAL_FLOWERS);
+		this.copy(BotaniaTags.Blocks.MISC_FLOATING_FLOWERS, BotaniaTags.Items.MISC_FLOATING_FLOWERS);
+		this.copy(BotaniaTags.Blocks.GENERATING_FLOATING_FLOWERS, BotaniaTags.Items.GENERATING_FLOATING_FLOWERS);
+		this.copy(BotaniaTags.Blocks.FUNCTIONAL_FLOATING_FLOWERS, BotaniaTags.Items.FUNCTIONAL_FLOATING_FLOWERS);
 		this.copy(BotaniaTags.Blocks.SPECIAL_FLOWERS, BotaniaTags.Items.SPECIAL_FLOWERS);
 		this.copy(BotaniaTags.Blocks.MINI_FLOWERS, BotaniaTags.Items.MINI_FLOWERS);
 		this.tag(BotaniaTags.Items.CONTRIBUTOR_HEADFLOWERS)